ADD [Proyecto 1B], log in and Sign Up with the two principal actors (Admin and Cajero), add 7 jsp to view the Implementation of the 2 iterations Inventario Produto and Inventario Descuentos with CRUD function, by last, TDD ProductoDAO and DescuentoDAO. Observations, changes were made to the database to meet the incremental objectives and visualize the two complete iterations

+++ src/main/java/servlets/incrementosProyecto/EliminarProductoServlet.java
@@ -1,4 +1,4 @@
-package servlets;
+package servlets.incrementosProyecto;
 
 import dao.ProductoDAO;
 import entity.Producto;
@@ -12,18 +12,21 @@ import java.io.IOException;
 
 @WebServlet("/eliminarProducto")
 public class EliminarProductoServlet extends HttpServlet {
+    private ProductoDAO productoDAO = new ProductoDAO();
 
     @Override
     protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        // Obtener el ID del producto a eliminar
         int id = Integer.parseInt(request.getParameter("id"));
 
-        // Crear un objeto Producto con el ID
-        Producto producto = new Producto();
-        producto.setId(id);
+        Producto producto = productoDAO.obtenerProductoPorId(id);
+        if (producto != null) {
+            productoDAO.eliminarProducto(producto);
 
-        // Eliminar el producto de la base de datos
-        ProductoDAO productoDAO = new ProductoDAO();
-        productoDAO.eliminarProducto(producto);
+            // Redirigir con mensaje de éxito
+            response.sendRedirect("formularioEliminarProducto.jsp?id=" + id + "&mensaje=eliminacionExitosa");
+        } else {
+            // Redirigir con mensaje de error
+            response.sendRedirect("formularioEliminarProducto.jsp?id=" + id + "&mensaje=errorEliminacion");
+        }
     }
 }
